requirements.txt
fastapi
uvicorn[standard]
chromadb
pandas
tqdm
python-dotenv
requests
streamlit
pydantic
numpy
